First semi-polished release
This first public version features a working clone of the Sirbu
model, with the addition of agents with fixed probability of
agreement, the encapsulation of all the mechanics of the model
into a single class (and its derivatives) and a script to quickly
feed condor multiple simulation with different parameters.

The code in the main programs and in the class file is more or
less polished, and it should be readable. The functions.py file
is still unpolished, because most of it could, and it will, be
removed.